+++ barista.php
@@ -10,7 +10,7 @@
  * Author:            WP Busters
  * Author URI:        https://wpbusters.com/
  *
- * Version:           0.6.37
+ * Version:           0.6.38
  * Requires at least: 5.8
  * Tested up to:      5.9
  * Requires PHP:      7.1
@@ -25,7 +25,7 @@ namespace Barista;
 define( 'BARISTA_PATH', \plugin_dir_path( __FILE__ ) );
 define( 'BARISTA_URL', \plugins_url( '/', __FILE__ ) );
 define( 'BARISTA_PLUGIN_FILE', __FILE__ );
-define( 'BARISTA_VERSION', '0.6.37' );
+define( 'BARISTA_VERSION', '0.6.38' );
 
 // define( 'BARISTA_DEMO', 'animate-up' ); phpcs:ignore Squiz.PHP.CommentedOutCode.Found.
 
@@ -33,34 +33,22 @@ if ( ! defined( 'BARISTA_DEVELOPMENT' ) ) {
 	define( 'BARISTA_DEVELOPMENT', false );
 }
 
-
-
 define( 'BARISTA_COMMAND_PRIORITY_WP_DASHBOARD', 1000 );
 define( 'BARISTA_COMMAND_PRIORITY_FEATURES', 2000 );
 define( 'BARISTA_COMMAND_PRIORITY_ACTIONS', 3000 );
 
 require __DIR__ . '/vendor/autoload.php';
+require __DIR__ . '/classes/class-singleton.php';
+require __DIR__ . '/classes/class-option.php';
+require __DIR__ . '/classes/class-settings.php';
 require __DIR__ . '/classes/class-collection.php';
+require __DIR__ . '/classes/class-recently-edited-posts.php';
 require __DIR__ . '/inc/updating.php';
 require __DIR__ . '/inc/admin.php';
 require __DIR__ . '/inc/frontend.php';
 require __DIR__ . '/inc/assets.php';
 require __DIR__ . '/inc/ajax.php';
-require __DIR__ . '/inc/actions/core.php';
-require __DIR__ . '/inc/actions/woocommerce.php';
-require __DIR__ . '/inc/actions/barista-settings.php';
-require __DIR__ . '/inc/actions/plugins.php';
-require __DIR__ . '/inc/actions/admin-menu.php';
-require __DIR__ . '/inc/actions/flush-rewrite-rules.php';
-require __DIR__ . '/inc/actions/flush-cache.php';
-require __DIR__ . '/inc/actions/recently-edited-posts.php';
-
-if ( BARISTA_DEVELOPMENT ) {
-	include __DIR__ . '/inc/actions/test-nested.php';
-}
-
-require __DIR__ . '/inc/actions/post.php';
-require __DIR__ . '/inc/actions/search.php';
+require __DIR__ . '/inc/init-commands.php';
 
 add_action( 'admin_init', __NAMESPACE__ . '\\admin_init' );