+++ classes/class-settings.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * Class settings.
+ *
+ * @package barista
+ */
+
+declare(strict_types=1);
+
+namespace Barista;
+
+/**
+ * Settings class.
+ */
+class Settings extends Option {
+	/**
+	 * Option name.
+	 *
+	 * @var string
+	 */
+	protected $option_name = 'settings';
+}